refactor: 중복 코드 hook으로 분리
